Update tests and all scripts to use adam_core orbits, remove now unnecessary tests and objects (like old Orbit class)
