--- doc/examples/mpc_orbits.csv
@@ -1,27 +1,0 @@
-,0,1,2,3,4,5,6,7,8,9,10,11
-0,0.0,-0.4390933617131744,-0.1257589780924159,-0.02090299118000981,0.007484308933440557,-0.02549827156943596,-0.007447399460385994,1.0,58000.00080074087,3.0,20.0,0.15
-1,1.0,-0.3140216978892445,0.7137191685860422,0.06395700844471017,-0.01699821554531536,-0.001621177958147811,0.008098761087529415,1.0,58000.00080074087,3.0,20.0,0.15
-2,2.0,1.069660168568243,-0.1091810313423533,-0.4009614224600879,-0.0005738733871607444,0.0148045888443323,-0.0004206593413873442,1.0,58000.00080074087,3.0,20.0,0.15
-3,3.0,1.146649218653312,0.4442347829093816,-0.4276439308423901,-0.009198907864594758,0.008522663718171699,0.0008579153396897662,1.0,58000.00080074087,3.0,20.0,0.15
-4,4.0,0.5571008932749256,-0.8515709683302697,0.01196777287922576,0.01593508296437815,0.005772078192645158,0.0004634885103096045,1.0,58000.00080074087,3.0,20.0,0.15
-5,5.0,-0.8195087914735473,-1.024391922126367,-0.06820864117682343,0.007579609349959913,-0.01066940860918116,-0.002172679692088396,1.0,58000.00080074087,3.0,20.0,0.15
-6,6.0,-0.09140722072579321,-1.304340688869913,0.2741461789883116,0.01562994084952533,-0.00661272020286746,0.0008803689566325372,1.0,58000.00080074087,3.0,20.0,0.15
-7,7.0,-1.069158496380121,-0.9018386016012201,-0.2661362048312304,0.006704643519994959,-0.0129471981145715,-0.0003371880510101042,1.0,58000.00080074087,3.0,20.0,0.15
-8,8.0,-0.2267100232186804,-2.037312055930078,0.003269521485770212,0.0108623181268777,0.004126258121972577,0.0003858249704382101,1.0,58000.00080074087,3.0,20.0,0.15
-9,9.0,1.085621272427007,-1.342298537106945,0.5177660223800957,0.01065676095419773,0.007179039083898097,-0.003324787815493938,1.0,58000.00080074087,3.0,20.0,0.15
-10,10.0,-0.9036255407528967,-1.52488584735319,-0.6862632415848007,0.01015087198957799,-0.007381596681530725,0.00178728023147973,1.0,58000.00080074087,3.0,20.0,0.15
-11,11.0,2.067010469927725,0.3429732056042801,0.1871156125963362,-0.001592411278119252,0.0103610603398434,0.00429102958758242,1.0,58000.00080074087,3.0,20.0,0.15
-12,12.0,2.325410505126934,0.9929926469902168,-0.8808980033143772,-0.007229353992222255,0.006751242223482295,-0.004059104051836333,1.0,58000.00080074087,3.0,20.0,0.15
-13,13.0,0.6310321043050902,-2.24519766267579,0.3336670997690477,0.009951048611696236,0.004794813331825265,-0.002676159128662144,1.0,58000.00080074087,3.0,20.0,0.15
-14,14.0,2.054661041199912,-0.8341004485101741,0.6196136200435236,0.005337871392227174,0.009572928979933995,0.003583443051098334,1.0,58000.00080074087,3.0,20.0,0.15
-15,15.0,-0.08262268266665583,2.708605953323662,0.2083056365711227,-0.01038047916629172,0.001069231556224658,-0.0001140263015168503,1.0,58000.00080074087,3.0,20.0,0.15
-16,16.0,0.5335311371774398,-5.20763828217245,-1.847956813870701,0.006972911905021999,0.0005713678801120065,0.001253934431033356,1.0,58000.00080074087,3.0,20.0,0.15
-17,17.0,-0.270716937713896,-5.697731135093401,0.2248897841083897,0.00685856475919825,-0.0004405555389105999,0.0002662305818316324,1.0,58000.00080074087,3.0,20.0,0.15
-18,18.0,-4.802147907178401,2.48760347016128,-1.613376836303095,-0.002763913097240356,-0.006365036430993691,-2.893763590972715e-05,1.0,58000.00080074087,3.0,20.0,0.15
-19,19.0,-4.594721757755073,3.343879458877429,0.4211596817284261,-0.00289153299252306,-0.005398778424685642,0.003113427532336405,1.0,58000.00080074087,3.0,20.0,0.15
-20,20.0,-0.9140604889908042,-26.70944214136232,6.380966421342785,0.002117924788461343,-0.001511770981078247,-0.0005076470050950094,1.0,58000.00080074087,3.0,20.0,0.15
-21,21.0,15.87454246952539,-11.65488173948501,6.130496666580919,-0.0008619367138370063,0.001508492017669211,0.0008079554285548605,1.0,58000.00080074087,3.0,20.0,0.15
-22,22.0,16.76382961811518,11.84247137955577,-3.37736807488939,-0.001725340506413252,0.003256405869321199,-9.52071384401337e-05,1.0,58000.00080074087,3.0,20.0,0.15
-23,23.0,33.75859656008453,23.7926066793143,0.9214856608529605,-0.001502839215619354,0.002300399302135842,8.73020950006359e-05,1.0,58000.00080074087,3.0,20.0,0.15
-24,24.0,14.13495240901712,22.96576678899176,0.8170986245916138,-0.003353555621352967,0.001767061747362163,4.938946831916796e-05,1.0,58000.00080074087,3.0,20.0,0.15
-25,25.0,28.92366156486876,24.53892273696946,2.448884400589299,-0.001416436065915651,0.002449714364547848,0.0002085644466749427,1.0,58000.00080074087,3.0,20.0,0.15
