+++ precovery/__init__.py
@@ -1,2 +1,1 @@
 from .main import precover  # noqa: F401
-from .utils import candidates_to_dataframe, frames_to_dataframe  # noqa: F401
