+++ precovery/_version.py
@@ -1,1 +1,1 @@
-__version__ = "0.2.dev189+ga42c2db.d20240918"
+__version__ = "0.2.dev189+g4b8d8ea.d20240919"
