+++ tests/test_benchmarks.py
@@ -1,7 +1,9 @@
 import pytest
+from adam_core.dynamics.propagation import propagate_2body
+from adam_core.propagator.adam_assist import ASSISTPropagator
+from adam_core.time import Timestamp
 
 from precovery.observation import Observation
-from precovery.orbit import PropagationMethod
 
 from .testutils import make_sourceframe_with_observations
 
@@ -49,12 +51,15 @@ def test_benchmark_store_observations(benchmark, frame_db, nobs):
 @pytest.mark.benchmark(group="orbits")
 @pytest.mark.parametrize("propagate_distance", [1, 10, 100, 1000])
 def test_benchmark_propagate_orbit_nbody(benchmark, sample_orbits, propagate_distance):
+
+    propagator = ASSISTPropagator()
     orbit = sample_orbits[0]
 
     def benchmark_case():
-        orbit.propagate(
-            [orbit._epoch + propagate_distance], method=PropagationMethod.N_BODY
+        times = Timestamp.from_mjd(
+            [orbit.coordinates.time.mjd()[0].as_py() + propagate_distance], scale="utc"
         )
+        propagator.propagate_orbits(orbit, times)
 
     benchmark(benchmark_case)
 
@@ -62,11 +67,13 @@ def test_benchmark_propagate_orbit_nbody(benchmark, sample_orbits, propagate_dis
 @pytest.mark.benchmark(group="orbits")
 @pytest.mark.parametrize("propagate_distance", [1, 10, 100, 1000])
 def test_benchmark_propagate_orbit_2body(benchmark, sample_orbits, propagate_distance):
+
     orbit = sample_orbits[0]
 
     def benchmark_case():
-        orbit.propagate(
-            [orbit._epoch + propagate_distance], method=PropagationMethod.TWO_BODY
+        times = Timestamp.from_mjd(
+            [orbit.coordinates.time.mjd()[0].as_py() + propagate_distance], scale="utc"
         )
+        propagate_2body(orbit, times)
 
     benchmark(benchmark_case)
